fix(banViolation.js): getLogStores import
