+++ api/test/__mocks__/KeyvMongo.js
@@ -28,42 +28,3 @@ jest.mock('@keyv/mongo', () => {
 
   return KeyvMongo;
 });
-
-// Mocking the getLogStores function
-jest.mock('../../cache', () => {
-  return {
-    getLogStores: jest.fn().mockImplementation(() => {
-      const EventEmitter = require('events');
-      const math = require('../../server/utils/math');
-      const mockGet = jest.fn();
-      const mockSet = jest.fn();
-      class KeyvMongo extends EventEmitter {
-        constructor(url = 'mongodb://127.0.0.1:27017', options) {
-          super();
-          this.ttlSupport = false;
-          url = url ?? {};
-          if (typeof url === 'string') {
-            url = { url };
-          }
-          if (url.uri) {
-            url = { url: url.uri, ...url };
-          }
-          this.opts = {
-            url,
-            collection: 'keyv',
-            ...url,
-            ...options,
-          };
-        }
-
-        get = mockGet;
-        set = mockSet;
-      }
-
-      return new KeyvMongo('', {
-        namespace: 'bans',
-        ttl: math(process.env.BAN_DURATION, 0),
-      });
-    }),
-  };
-});
